requirements.txt
# ---------------- Core App ----------------
streamlit>=1.38.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.5.0
matplotlib>=3.8.0
seaborn>=0.13.0

# ---------------- Model & Evaluation ----------------
torch>=2.2.0
transformers>=4.43.0
accelerate>=0.32.0
evaluate>=0.4.2

# ---------------- LangChain & Orchestration ----------------
langchain>=0.2.12
langchain-community>=0.2.11
langchain-core>=0.2.29
langchain-experimental>=0.0.64


# ---------------- Optional: Viz & Metrics ----------------
plotly>=5.22.0
scipy>=1.12.0

# ---------------- Dev & Utils ----------------
python-dotenv>=1.0.1
tqdm>=4.66

reportlab

